Correctly handle truncation of content and trailing field sections
RFC-9292

3.8. Padding and Truncation

Decoders MUST treat missing truncated fields as equivalent to having
been sent with the length field set to zero.

5.1. Request Example

Known-Length Binary Encoding of Request:

00034745 54056874 74707300 0a2f6865  ..GET.https../he
6c6c6f2e 74787440 6c0a7573 65722d61  llo.txt@l.user-a
67656e74 34637572 6c2f372e 31362e33  gent4curl/7.16.3
206c6962 6375726c 2f372e31 362e3320   libcurl/7.16.3
4f70656e 53534c2f 302e392e 376c207a  OpenSSL/0.9.7l z
6c69622f 312e322e 3304686f 73740f77  lib/1.2.3.host.w
77772e65 78616d70 6c652e63 6f6d0f61  ww.example.com.a
63636570 742d6c61 6e677561 67650665  ccept-language.e
6e2c206d 690000                      n, mi..

The message can be truncated with no effect on interpretation.
In this case, the last two bytes -- corresponding to content and
a trailer section -- can each be removed without altering the semantics
of the message.

Indeterminate-Length Binary Encoding of Request:

02034745 54056874 74707300 0a2f6865  ..GET.https../he
6c6c6f2e 7478740a 75736572 2d616765  llo.txt.user-age
6e743463 75726c2f 372e3136 2e33206c  nt4curl/7.16.3 l
69626375 726c2f37 2e31362e 33204f70  ibcurl/7.16.3 Op
656e5353 4c2f302e 392e376c 207a6c69  enSSL/0.9.7l zli
622f312e 322e3304 686f7374 0f777777  b/1.2.3.host.www
2e657861 6d706c65 2e636f6d 0f616363  .example.com.acc
6570742d 6c616e67 75616765 06656e2c  ept-language.en,
206d6900 00000000 00000000 00000000   mi.............

This indeterminate-length encoding contains 10 bytes of padding.
As two additional bytes can be truncated in the same way
as the known-length example, anything up to 12 bytes can be removed
from this message without affecting its meaning.